Consolidate exchange related error types
