+++ academy/agent.py
@@ -14,11 +14,11 @@ from typing import TypeVar
 from academy.behavior import BehaviorT
 from academy.context import ActionContext
 from academy.context import AgentContext
-from academy.exception import BadEntityIdError
-from academy.exception import MailboxClosedError
 from academy.exception import raise_exceptions
 from academy.exchange import AgentExchangeClient
 from academy.exchange import ExchangeFactory
+from academy.exchange.exception import ExchangeError
+from academy.exchange.exception import MailboxClosedError
 from academy.exchange.transport import AgentRegistrationT
 from academy.exchange.transport import ExchangeTransportT
 from academy.handle import Handle
@@ -143,11 +143,16 @@ class Agent(Generic[BehaviorT], NoPickleMixin):
         assert self._exchange_client is not None
         try:
             await self._exchange_client.send(response)
-        except (BadEntityIdError, MailboxClosedError):  # pragma: no cover
+        except MailboxClosedError:  # pragma: no cover
             logger.warning(
-                'Failed to send response from %s to %s. '
-                'This likely means the destination mailbox was '
-                'removed from the exchange.',
+                'Failed to send response from %s to %s because the '
+                'destination mailbox was terminated.',
+                self.agent_id,
+                response.dest,
+            )
+        except ExchangeError:  # pragma: no cover
+            logger.exception(
+                'Failed to send response from %s to %s.',
                 self.agent_id,
                 response.dest,
             )
