+++ academy/handle.py
@@ -28,7 +28,7 @@ else:  # pragma: <3.11 cover
 
 from academy.exception import HandleClosedError
 from academy.exception import HandleNotBoundError
-from academy.exception import MailboxClosedError
+from academy.exchange.exception import MailboxClosedError
 from academy.identifier import AgentId
 from academy.identifier import EntityId
 from academy.identifier import UserId
